Smart Scan/requirements.txt
flask
google-generativeai
python-dotenv
pillow
